requirements.txt
allure-pytest
pytest
pytest-base-url==1.4.2
requests
python-dotenv
pytest-xdist
requests
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
asserts==0.13.1